// frontend/src/types/langgraph.ts
export type AssistantSortBy = "assistant_id" | "graph_id" | "name" | "created_at" | "updated_at";
export type ThreadSortBy = "thread_id" | "status" | "created_at" | "updated_at";
export type SortOrder = "asc" | "desc";

export type ThreadStatus = "idle" | "busy" | "interrupted" | "error";
export type RunStatus = "pending" | "running" | "error" | "success" | "timeout" | "interrupted";

export interface Assistant {
    /** The ID of the assistant. */
    assistant_id: string;
    /** The ID of the graph. */
    graph_id: string;
    /** The assistant config. */
    config: Config;
    /** The time the assistant was created. */
    created_at: string;
    /** The assistant metadata. */
    metadata: Metadata;
    /** The version of the assistant. */
    version: number;
    /** The name of the assistant */
    name: string;
    /** The description of the assistant */
    description?: string;
    /** The last time the assistant was updated. */
    updated_at: string;
}
// ...
export interface SearchAssistant {
    graphId?: string;
    metadata?: Metadata;
    limit?: number;
    offset?: number;
    sortBy?: AssistantSortBy;
    sortOrder?: SortOrder;
}

export interface CreateThread {
    metadata?: {
        [key: string]: any;
    };
    threadId?: string;
    ifExists?: "raise" | "do_nothing";
    graphId?: string;
    supersteps?: [];
}

export interface SearchThread {
    metadata?: {
        [key: string]: any;
    };
    limit?: number;
    offset?: number;
    status?: "idle" | "busy" | "interrupted" | "error";
    sortBy?: "thread_id" | "status" | "created_at" | "updated_at";
    sortOrder?: "asc" | "desc";
}